+++ hydromt/model/components/mesh.py
@@ -9,8 +9,8 @@ import xugrid as xu
 from pyproj import CRS
 from shapely.geometry import box
 
-from hydromt._io.readers import _read_ncs
 from hydromt.gis.raster import GEO_MAP_COORD
+from hydromt.io.readers import read_ncs
 from hydromt.model.components.base import ModelComponent
 from hydromt.model.components.spatial import SpatialModelComponent
 from hydromt.model.steps import hydromt_step
@@ -181,7 +181,7 @@ class MeshComponent(SpatialModelComponent):
         self._initialize(skip_read=True)
 
         filename = filename or str(self._filename)
-        files = _read_ncs(
+        files = read_ncs(
             filename,
             root=self.root.path,
             single_var_as_array=False,